server/pyproject.toml
[project]
name = "trustmodel-server"
version = "0.1.0"
description = "TrustModel Agent Eval - Backend Server"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "TrustModel Team" }
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "redis>=5.0.1",
    "celery>=5.3.6",
    "pynacl>=1.5.0",
    "anthropic>=0.18.0",
    "websockets>=12.0",
    "structlog>=24.1.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "tenacity>=8.2.3",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-passlib>=1.7.7",
    "types-python-jose>=3.3.4",
    "types-redis>=4.6.0",
    "types-pyyaml>=6.0.12",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
